Added vault and roll start up and shutdown for gucumber
