+++ .gitignore
@@ -1,4 +1,4 @@
-internal/runroll/runroll
+internal/runutils/runroll/runroll
 echo/echo
 cbserver/cbserver
 rollsvcs/rollsvcs
