+++ internal/features/integrationtest/integrationtest.feature
@@ -1,0 +1,5 @@
+Feature: Integration Tests
+
+  Scenario: Pocess Management
+    Given Some tests to run
+    Then The test process and supporting containers are started before and stopped after
