+++ internal/runutils/runroll/runroll.go
@@ -1,0 +1,9 @@
+package main
+import (
+	"github.com/xtraclabs/roll/internal/runutils"
+)
+
+func main() {
+	shutdownDone := runutils.RunVaultAndRoll()
+	<-shutdownDone
+}
